add rpcFileData arags

+++ eth/protocols/eth/handler.go
@@ -219,7 +219,7 @@ func handleMessage(backend Backend, peer *Peer) error {
 	if err != nil {
 		return err
 	}
-	if msg.Size > maxMessageSize {
+	if uint64(msg.Size) > maxMessageSize {
 		return fmt.Errorf("%w: %v > %v", errMsgTooLarge, msg.Size, maxMessageSize)
 	}
 	defer msg.Discard()
